+++ .env
@@ -1,0 +1,7 @@
+PORT=8080
+APP_ENV=local
+DATABASE_PORT=3306
+DATABASE_NAME=blueprint
+DATABASE_USERNAME=melkey
+DATABASE_PASSWORD=password1234
+DATABASE_ROOT_PASSWORD=password4321
