+++ app/internal/database/database.go
@@ -1,0 +1,46 @@
+package database
+
+import (
+	"database/sql"
+	"fmt"
+	"log"
+	"time"
+
+	_ "github.com/go-sql-driver/mysql"
+)
+
+type DatabaseConfig struct {
+	Host     string
+	Port     int
+	User     string
+	Password string
+	Database string
+}
+
+func InitializeDatabase(cfg DatabaseConfig) (*sql.DB, error) {
+	dsn := fmt.Sprintf("%s:%s@tcp(%s:%d)/%s?parseTime=true&charset=utf8mb4&collation=utf8mb4_unicode_ci",
+		cfg.User,
+		cfg.Password,
+		cfg.Host,
+		cfg.Port,
+		cfg.Database,
+	)
+
+	db, err := sql.Open("mysql", dsn)
+
+	if err != nil {
+		return nil, fmt.Errorf("failed to open database connection; %w", err)
+	}
+
+	db.SetMaxOpenConns(25)
+	db.SetConnMaxIdleTime(25)
+	db.SetConnMaxLifetime(5 * time.Minute)
+
+	if err := db.Ping(); err != nil {
+		return nil, fmt.Errorf("failed to ping database: %w", err)
+	}
+
+	log.Println("Connection to Database established successfully.")
+
+	return db, nil
+}
