--- go.mod
@@ -1,14 +1,0 @@
-module app
-
-go 1.21.5
-
-require (
-	github.com/go-sql-driver/mysql v1.8.1
-	github.com/google/uuid v1.6.0
-)
-
-require (
-	filippo.io/edwards25519 v1.1.0 // indirect
-	golang.org/x/crypto v0.29.0
-	golang.org/x/sys v0.27.0 // indirect
-)
